selinux: (re)start fails to read /var/lib/nfs when it's a symlink
In Red Hat Gluster Storage, the HA configuration changes /var/lib/nfs to
be a symlink to a shared state directory. New selinux policy is prohibiting
the reading of that symlink and ganesha is prevented from starting when it
tries to read it.

Also fixes another AVC (cluster_t:dbus send_msg) seen at the same time.

Change-Id: I567fb165b0e8077f4071b0dc75fa331d8dc42ee0
Signed-off-by: Kaleb S. KEITHLEY <kkeithle@redhat.com>