Rework fsal_filesystem handling for better config reload and handle validation
There were some problems in claiming and unclaiming filesystems during
config reload that have been fixed.

Also, it is now possible for a FSAL to check that the fsal_filesystem
a handle belongs to is actually exported by the export from the handle.
This reduces the security exposure to handle spoofing. Note that the
entirety of a filesystem that is exported by multiple exports is
accessible via any of the exports, so for the best export security
each filesystem should only be exported by a single export and should
be exported in it's entirety.

Change-Id: If571e29054d639ad0dd236b4e4fa5b91f54ce426
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>