Start defining Darwin build based on FreeBSD
Some of the FreeBSD build logic as well as OS-specific source and
headers can be used on macOS. So move those files to a generic BSD
category, and define a BSDBASED build type and use that where it makes
sense. I may have missed some places where it makes sense to use
"#if BSDBASED" rather than "#if FREEBSD", but these are the obvious ones
that are necessary for my macOS build to function.

The subr.c and mntent compatibility code are not completely
macOS-compatible yet. I will adjust them later.

Signed-off-by: Matthew DeVore <matvore@google.com>
Change-Id: Idec2f7ac960016c15774490981d2ea21dd2b62dd