RADOS_KV: make rados_kv and rados_ng respect nodeid setting
Proper interaction with FSAL_CEPH requires that when one server takes
over for another, that it use the same nodeid setting as the original
server. This is properly implemented in rados_cluster backend, but
not in rados_ng or rados_kv.

Implement a get_nodeid op for rados_ng and rados_kv that just clones
the rados_kv_param value if there is one.

Change-Id: I9451e79c3ea7f051062bcad093865f7744c76151
Reported-by: Lars Kellogg-Stedman <lars@redhat.com>
Signed-off-by: Jeff Layton <jlayton@redhat.com>