Get time_delta from FSAL dynamic fsinfo
Note that this change by itself doesn't do much. Most in-tree FSALs
report 1 second time_delta...

FSAL_PROXY_V3 does fetch this attribute from the server though it
should be noted the Linux kernel server hard codes this to 1 second
for NFSv3... Except it actually gets invarsec from FSSTAT which is
only in seconds so is not useful...

Fill out ALL of dynamiinfo in encode_fetch_fsinfo if there is no
object. This is unlikely to actually be relevant but best to be complete.

Change-Id: I259eaa032fcc90e5e4d9b90d0ec7a3c3e9231ef1
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>