Change from default of 1 second for time_delta to 100 ns.
Set the default time_delta value in a constant to make it easy to
change.

There doesn't seem to be a good way of getting a true value for
this, even with the PROXY FSALs (PROXY_V3 was using invarsec from
FSSTAT which is only seconds and not ns, and PROXY_V4 for sure isn't
decoding time_delta and may not be asking for it). The previous
global default of 1 second is causing issues with at least one
client throwing scary sounding warning messages and maybe changing
behavior.

We chose 100 ns as a "reasonable" default that doesn't overstate
the resolution but also doens't make us look bad in some way. There
is reasonable arguments for anything from 1 ns to 1 ms.

Interestingly FSAL_KVSFS doesn't fill ANYTHING in for get_dynamic_info
which probably is not good but I don't know anything about kvfsf to
have any idea what to do there but a todo comment seems in order.

Change-Id: Ibf43dfe9d970b908bf05ef699d45fe3bcafc3b7b
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>