Fixed NFSv4 lock reclaim issue in failover scenario
As per the RFC#5661, in the cluster environment, the server
should return server_scope string which is common across all
the NFS server in the cluster.The NFS client will use server_scope
to reclaim locks if there is a failover from one server to another.

Currently, we are returing hostname as server_scope (in exchange id
request) which is wrong.

Added new config parameter server_scope in ganesha.main.conf and it is
settable value. The preferred value for server_scope is clustername.

Change-Id: I1ce49bac3ada0e9166491fbfce8528448ad279c2
Signed-off-by: Prabhu Murugesan <prabhu.murugeshan@ibm.com>