Remove ambiguous use of "export" in header file
fsal_commonlib.h cannot be easily included from a C++ source file
because it uses "export" as a parameter name, which is a C++ keyword.

This is the only instance of "export" causing problems for me at the
moment. I didn't do a global cleanup as this would unnecessarily change
headers that are not part of Ganesha's public API.

Signed-off-by: Matthew DeVore <matvore@google.com>
Change-Id: I72b95e2124a90c82438a30457fc092e3cf7a5c32