MDCACHE - Ensure ATTR4_FS_LOCATIONS and ATTR4_SEC_LABEL bits aren't lost
The refresh/copy/ref-handoff code depends on the ATTR4_* bits in
request_mask always matching the associated pointers.  However,
mdcache_refresh_attrs() overwrites request_mask with the given one,
losing this relationship.  Fix it up, to make sure the bits are always
set when the pointers are non-NULL.  This was already done for ACL, but
not for FS_LOCATIONS or SEC_LABEL

Change-Id: I5784a797545f0ab4bf8562594431134484fbe757
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>