WIP - Synchonizing I/O and file open/close/reopen without holding rwlock
For async, we will need to protect I/O from open/close/reopen with the
operation being completed on a different thread. Because of this, we
can't use a pthread_rwlock, so we basically implement our own.

This also allows reducing the contentioon for the obj_lock since we
actually just lock the fsal_fd we are using.

This implementation ALSO acquires temporary share reservations for
more operations. The impact of this actually is that instead of share
operations blocking on I/O, they may instead fail.

Change-Id: Ib36e112c5599f15ba0522750447175c30badc5a2
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>