checkpatch fixups
We're introducing a new version of checkpatch, plus it's been a while
since checkpatch has been run on the entire code source and stuff has
snuck in.

There are still some errors when running src/scripts/runcp.sh -c but
they are not really fixable but I don't want to ignore the particular
check because it may be in general relevant.

Change-Id: I74b42bf38082752f783b96420127f975a3507aee
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>