systemd: eliminate a (benign?) error when enabling nfs-ganesha.service
See https://bugzilla.redhat.com/show_bug.cgi?id=1907534 (private)

Enabling ganesha, e.g. with `systemctl enable nfs-ganesah` results in
the following error msg:

...
The unit files have no installation config (WantedBy, RequiredBy, Also, Alias
settings in the [Install] section, and DefaultInstance for template units).
This means they are not meant to be enabled using systemctl.
Possible reasons for having this kind of units are:
1) A unit may be statically enabled by being symlinked from another unit's
   .wants/ or .requires/ directory.
2) A unit's purpose may be to act as a helper for some other unit which has
   a requirement dependency on it.
3) A unit may be started when needed via activation (socket, path, timer,
   D-Bus, udev, scripted systemctl call, ...).
4) In case of template units, the unit is meant to be enabled with some
   instance name specified.

But the nfs-ganesha.service file does have an [Install] section, and it
was not apparent that the error is because, as explained by one of the
systemd devs, the nfs-ganesha-lock.service file does not have an [Install]
section.

Change-Id: Iea168b14feb4426e352b1ea88140b17aa47efb26
Signed-off-by: Kaleb S. KEITHLEY <kkeithle@redhat.com>