FSAL_VFS: Put debug ACL code under a single #if defined()
More code is coming here.  Having more #ifdef's sprinkled through the
function bodies is going to make this unreadable and unmaintainable.
So, create a single section of code to be used when
ENABLE_VFS_DEBUG_ACL is enabled.  Factor out the bits that are
reusable.

Use #if defined() for consistency with future use of #elif defined().

Signed-off-by: Martin Schwenke <mschwenke@ddn.com>
Change-Id: Ic9044e66cd034e1aeb9ec8aea59e17fb89a45b63