build: Add check for libacl
While this is useful on its own, it seems to be required for a
subsequent commit that uses check_symbol_exists().

Unfortunately, this triggers a developer warning about this policy:

  https://cmake.org/cmake/help/latest/policy/CMP0075.html#policy:CMP0075

This seems to suggest that it may no longer work with cmake 3.23.1.
My testing is with cmake 3.23.0 (Debian testing), 2.8.12.2 (CentOS 7)
and 3.22.2 (FreeBSD 13).

Signed-off-by: Martin Schwenke <mschwenke@ddn.com>
Change-Id: Icfbc76117298a60cb106950c80311d483251e9e9