STATE: Fix ganesha crash at nlm4_unlock() in blocked lock req
When a blocked lock is granted, we need to take ref of file
handle obj if the lock list is empty. The ref will be released
in unlock() if the lock list is empty. This will fix possible
crash in blocked lock request. With out this fix, the ganesha
will crash at nlm4_unlock() if a lock is granted to client after
lock conflict and the file is removed by another thread from
same client.

Change-Id: Idaee9e85574969de089950fac3e50ae736a170de
Signed-off-by: Prabhu Murugesan <prabhu.murugeshan@ibm.com>