Add Prometheus and Grafana monitoring stack.
Measure, export, aggregate and display monitoring metrics for:
1. NFS requests, aggregated per {operation} and per {operation, export}
- Request rates
- Sent / received throughput rates in bytes/second.
- Latency percentiles in ms as a histogram.
- Request and Response sizing in bytes as a histogram.
2. Metadata cache hit and miss rates,
   per {operation} and per {operation, export}
3. Rates of received and completed rpcs.
4. Number of RPCs in flight.
5. NFS client request and throughput rates.
6. Number of active NFS clients.

By default, export metrics on TCP port 9587. Reuse the same port that is
allocated to the nfs-ganesha exporter at
https://github.com/prometheus/prometheus/wiki/Default-port-allocations

Other changes:
- Add a parameter to NFS core parameters to allow the user change the
  monitoring port.
- As this change adds new library dependencies,
  guard it by the USE_MONITORING definition.
- Add sample configuration file for Prometheus metrics aggregation.
- Add sample configuration file for Grafana console.

This change adds the first C++ file to NFS Ganesha. I tried to implement
the monitoring in C only using the digitalocean C Prometheus client.
This turned out to be a bottleneck that wasted 25% of all cpu and
limited overall request throughput:
https://github.com/digitalocean/prometheus-client-c/issues/59
This flatlined Ganesha requests/second performance. As there are no
other C libraries, the monitoring code instead depends on a C++ library,
which doesn't have this performance problem. The Prometheus C++ library
is thread safe, so no locking is added in NFS Ganesha.

Signed-off-by: Bjorn Leffler <leffler@google.com>
Change-Id: I825cf2283f6e9e8c450cbf2923f92137fbee7e58