Protocols: Add support for read2 and write2 that ask for resume
Sometimes a FSAL needs to request a resume because it has to do more work
after an I/O completion callback that it can not do on the thread the
callback is running on.

Currently this is used to close a temporary fsal_fd but future uses might
add additional requirements.

Change-Id: I74f6dfc90be26524b9cf1f0ffac1b64757ca2bbd
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>