Protocols: Add support for read2 and write2 that ask for resume
Sometimes a FSAL needs to request a resume because it has to do more work
after an I/O completion callback that it can not do on the thread the
callback is running on.

The FSAL will have determined during the back end filesystem's call back
that additional action must be taken that must not be done on the call back
thread. By requesting a resume, the caller will re-schedule so the I/O
operation can be resumed on a caller thread, not on the back end
filesystem's call back thread.

The immediate case that this will be used for is when a temp_fd (see
the fsal_fd management) is being used for the operation. That temp_fd
must be closed when the I/O is complete, however, at least in the case
of cephfs, close can not be called on ceph's call back thread.

Note that 9P doesn't currently make non-blocking async I/O calls, it
instead blocks the 9P worker thread until completion, all of which is
handled in the fsal_helper functions and thus those functions will also
handle the need to resume the read2 or write2 call.

Change-Id: I74f6dfc90be26524b9cf1f0ffac1b64757ca2bbd
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>