FSAL: Add global fd LRU to FSAL layer
Eventually the old fd reaping will be removed from MDCACHE.

In order to not complicate parameters, mdcache passes parameters to the
FSAL layer and invokes the startup and shutdown. Of course mdcache is
actually part of the FSAL layer so this isn't horrible... Note that we
don't have a global FSAL config block to move the config parameters into,
plus we would need to provide backward compatibility so this sort of
makes sense.

Change-Id: I8ae462cd9569b6acd5c44b7b0f74102a8b986f69
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>