Readdir performance optimization.
Readdir by chunk is slow if the number of file is large, and
Entries_HWMark of configuration is not large enough.
It fills all files to chunks in the directory under the current offset
at one time, but noly read a part of chunk in a readdir. The entries
cache in chunks may can't find in global avl which causes release the
chunk and refill chunks again.

Add the chunk dirty detection mechanism. If the chunk has set dirty, go
to the readdir uncached branch next readdir, and recover to the read
chunk after a period of time.

Change-Id: I743c81921f81bbd2677de11e50023a3c4f44c69a
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>