FSAL_PROXY_V3: Ensure that the used xid is unique
If the used XID is the same as other call that is identical, it may
cause the DRC to drop the request which cause the FSAL thread and client
to halt.
Using rand_r is not good enough as multiple threads calling it at the
same time (using same 'seedp') may produce the same number.

Signed-off-by: Assaf Yaari <assafy@google.com>
Change-Id: I725e62f91638ee39aad4c990cffaa4b470f0aa45