readdir: set L2 cache control water to improve readdir performance.
The mdcache cache management mechanism causes that if cache hits
continuously, most caches will be concentrated in L1 linked lists
over time, and L2 caches are severely insufficient. A readdir
corresponding package consists of multiple file/folders, Cache in
front of the response package may be obsoleted. In this case, if
the client does not receive a file/folder and trigger retransmission,
the server will releases the chunk the retransmitted file/folder
belongs and fill the chunk repeatedly if dir_chunk is not set to 0.
This severely degrades readdir performance.

The code provides a strategy to optimize cache management: during a
readdir op, the number of L2 list caches is  controlled above the
set water level to avoid the recently filled caches be obsoleted,
thus improving readdir performance.

Signed-off-by: iamtiger51 <1140520975@qq.com>
Change-Id: I46901ae86f74b2598fe0a6763956580e31cab5cd