readdir: change the readdir cache hit adjustment policy to improve readdir performance.
If the client not receive a file/floder from the last readdir, it will retransfers
missed file/folders. If  L2 caches are severely insufficient, some cache are
eliminated and new cache will insert into L2, the cache not eliminated will
moved to L1 head from L2 . Because the L2 cache does not meet the condition
of being eliminated at this time, the cache in the L1 header will be eliminated
soon. When retransmission occurs again, this behavior can easily result in
partial cache obsolescence of the chunk, resulting in repeated release and
filling of the chunk.

The code provides a strategy to optimize cache management: if a cache hit is
caused by readdir, the cache is no longer moved to the head of L1, but to the
tail of L2, preventing the hit cache from being eliminated immediately.

Signed-off-by: iamtiger51 <1140520975@qq.com>
Change-Id: Ibcacc97f1bf2731cad3999ca8a03bd950c81e9e9