Expire lease immediately when reserve_lease fails.
If a client request bumps into a lease that should have expired,
expire it immediately rather than letting the reaper thread
clean it up. This is because the client will set up a new clientid
(EXCHANGE_ID or SET_CLIENTID) which ends up having the effect of
not actually expiring the state.

Change-Id: I61bfc3fad849592b505af8047ca96af2d0b14085
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>