Handle suid, sgid bits reset during chown operation
fsal_helper: Change order of mode bits check in fsal_setattr

Previously, the first check set ATTR_MODE as valid if ATTR_OWNER or
ATTR_GROUP was valid. Then, the the next check looked at if ATTR_MODE
was valid. But, there could be a case where it is not originally valid
and just because valid due to our code.

Changing the order fixes this issue.

Change-Id: I7c80f03c77f062ce5d8f4ef89b5e326dd18310df
Signed-off-by: Lior Suliman <liorsu@gmail.com>