Donot provide delegations when another client has a write open file handle
Bug description: Today we provide read delegations
to clients even when another client has a write open file handle.
This causes read delegated clients to hold stale data(but
think that the data is valid) when the write client
tries to write to the file.

Experiments show that when the write client writes to the file
we respond to the write client saying write is done and then do
a delegation recall much later this is problematic.

How this bug was found: This bug was found by running pynfs
delegation tests against ganesha server and comparing it
against linux kernel nfs server.

Fix: This patch fixes that issue by introducing another
counter to keep track of the number of write opens. We don't provide
delegations if another client has opened this file in write mode

Change-Id: Icc203c9c144097efaed2b4fa928abe6d4ad6099e
Signed-off-by: Deepak Arumugam Sankara Subramanian <deepakarumugam.s@nutanix.com>