build: Fix Python detection
Commit c43049fc42b0a1d39b48ddbf93af56223ba5f9ee introduced copies of
several cmake modules, which are now breaking Python detection.  The
commit message does not give details of benefits of these versions of
the modules.  It is unclear if they were different to the upstream
cmake modules at the time and this can't be resolved without an
archaeological dig.  Whatever the case, the modules have almost
certainly been improved upstream to work around the reason for their
addition here.

Just delete the duplicate modules introduced by the above commit.

This leaves src/cmake/modules/FindGTest.cmake as the only apparent
duplicate module.

Note that this change introduces a developer warning:

CMake Warning (dev) at /usr/share/cmake-3.25/Modules/FindPackageHandleStandardArgs.cmake:438 (message):
  The package name passed to `find_package_handle_standard_args`
  (LSB_RELEASE) does not match the name of the calling package (LSB).  This
  can lead to problems in calling code that expects `find_package` result
  variables (e.g., `_FOUND`) to follow a certain pattern.
Call Stack (most recent call first):
  cmake/modules/FindLSB.cmake:85 (find_package_handle_standard_args)
  CMakeLists.txt:1632 (find_package)
This warning is for project developers.  Use -Wno-dev to suppress it.

Given that this is only a warning, this can be fixed in a subsequent
commit.

Signed-off-by: Martin Schwenke <mschwenke@ddn.com>
Change-Id: I0cb1fb4d3dd9674d52720095961ed6b56dda6c45