FSAL_MDCACHE/mdcache_handle: Fixed rename atomicity
This fixes a Time-of-check Time-of-use bug checking for the file handle of
the source before moving and then the file handle of the target before
moving and comparing them. There can be a scenario that two or more
clients will perform the rename at the same time both will pass the source
lookup and one of them will perform the move while the other can get a
success for it as well. This damaged the atomicity of the rename
operation. The FSAL should handle the case when those are different
entries with the same file handle (hard-links).This fixes a Time-of-check Time-of-use bug checking for the file handle of the source before moving and then the file handle of the target before moving and comparing them. There can be a scenario that two or more clients will perform the rename at the same time both will pass the source lookup and one of them will perform the move while the other can get a success for it as well. This damaged the atomicity of the rename operation. The FSAL should handle the case when those are different entries with the same file handle (hard-links).

Change-Id: I70bcfa28e9369c52d6b1385ac70eed20fc07f48d
Signed-off-by: Lior Suliman <liorsu@google.com>