FSAL_VFS: Don't attempt to get ACLs from symbolic links
acl_get_fd() returns NULL, failing with EBADF, if given a file
descriptor for a symlink opened with O_NOFOLLOW.

Add an early return with success.  This turns the function into a
no-op for symlinks, so should guarantee now resource leakage.

The GlusterFS FSAL handles this slightly differently.  It attempts to
get the ACLs and converts failure to success on certain errors.
However, here in the VFS FSAL, the errno for a failure of this type
might vary between filesystems - various manpages provide little
guidance - so avoiding the lookup is probably sane.

Signed-off-by: Martin Schwenke <mschwenke@ddn.com>
Change-Id: I69945978f25522b26e312f3577e7a13e50a497f1