fsal.h: Don't unset the ATTR_ACL from valid_mask in case it was set by src
In fsal_copy_attrs, it might be that the source attrs as was returned by
the FSAL, can indicate that the ACL info is valid but the acl field
points to NULL when there are no ACL entries stored in the BE.
We don't want in this case (which I believe is the common case) to
allocate new fsal_acl_data_t and call nfs4_acl_new_entry just for marking
zero entries.

The unset to the flag caused Ganesha to return malformed response to
GETATTR that contained the ACL mask.
Note that encode_acl do support the case where args->attrs->acl is NULL.

Signed-off-by: Assaf Yaari <assafy@google.com>
Change-Id: I2c901d2f65c5e632c60dd2c5fcd25a98fe519c69