Idmapper: Limit max user-cache entries, and add a fifo queue
This change limits the max number of user entries in the cache.
The limit is configurable through the DIRECTORY_SERVICES
section in Ganesha config.

The cache eviction strategy is determined using the fifo queue.
The fifo queue order mimics the order of expiration time of the
cache entries, since expiration time is a linear function of
the insertion time --

expiration_time = insertion_time + cache_validity (constant)

The head of the queue therefore contains the entry with least
time-validity, and the tail of the queue contains the entry
with most time-validity. The eviction happens from the head,
and insertion happens at the tail.

Change-Id: I78c4544f3b430b1d0be16de766ce5c1def3f79ba
Signed-off-by: Dipit Grover <dipit@google.com>