Idmapper: Introduce negative cache for user entries
This change introduces a negative cache for storing user entries
that failed idmapping. The validity and max-count of such entries
is derived from the DIRECTORY_SERVICES section of the Ganesha
config.

This cache is maintained separately from positive cache, to
allow for independent operation, configuration and maintenance.

Note that we do not store the anonymous uid and gid resolved for
such entries, because these resolved ids for a given entry can
vary across different incoming nfs op-requests. Each request
uses the anonymous uid,gid configured for the respective client's
export that is being accessed.

Change-Id: I4c2e085d33dbf70feb5d38bad5d1014669ec4953
Signed-off-by: Dipit Grover <dipit@google.com>